requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary>=2.9.0

# Notebook dependencies
xarray>=2023.1.0
netCDF4>=1.6.0
geopandas>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
shapely>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
jupyter>=1.0.0
ipykernel>=6.0.0
pyproj>=3.5.0
rasterio>=1.3.0
Pillow>=10.0.0
requests>=2.31.0
trino>=0.330.0
tqdm>=4.65.0
pyopensky>=2.0.0
